fix: revert drizzle and change method name
