+++ apps/aircall/drizzle/0001_fearless_molly_hayes.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "organisations" ADD COLUMN "auth_user_id" text NOT NULL;
