+++ apps/aircall/drizzle/meta/0001_snapshot.json
@@ -1,6 +1,6 @@
 {
-  "id": "9124f28b-c819-4e28-9c5f-2db7ddbc33b8",
-  "prevId": "00000000-0000-0000-0000-000000000000",
+  "id": "515f4648-d74a-4225-9093-2b0a3bc2e3e4",
+  "prevId": "9124f28b-c819-4e28-9c5f-2db7ddbc33b8",
   "version": "5",
   "dialect": "pg",
   "tables": {
@@ -22,6 +22,12 @@
         },
         "access_token": {
           "name": "access_token",
+          "type": "text",
+          "primaryKey": false,
+          "notNull": true
+        },
+        "auth_user_id": {
+          "name": "auth_user_id",
           "type": "text",
           "primaryKey": false,
           "notNull": true
