+++ apps/aircall/drizzle/meta/_journal.json
@@ -5,8 +5,15 @@
     {
       "idx": 0,
       "version": "5",
-      "when": 1724922025175,
-      "tag": "0000_exotic_jane_foster",
+      "when": 1715850338177,
+      "tag": "0000_smart_kitty_pryde",
+      "breakpoints": true
+    },
+    {
+      "idx": 1,
+      "version": "5",
+      "when": 1724924387762,
+      "tag": "0001_fearless_molly_hayes",
       "breakpoints": true
     }
   ]
